Move Zoltan implementation to core/zoltan.c
